Move helpers to ui folder
