+++ panel/src/components/Layout/ListItem.vue
@@ -43,7 +43,7 @@
 </template>
 
 <script>
-import previewThumb from "@/helpers/previewThumb.js";
+import previewThumb from "@/ui/helpers/previewThumb.js";
 
 export default {
   inheritAttrs: false,
