+++ panel/src/ui/helpers/clone.js
