+++ panel/src/ui/helpers/isComponent.js
