+++ panel/src/ui/helpers/isUploadEvent.js
