+++ panel/src/ui/helpers/previewThumb.js
