+++ panel/src/ui/helpers/ratio.js
