+++ panel/src/ui/helpers/slug.js
