+++ panel/src/ui/helpers/string.js
