+++ panel/src/ui/helpers/upload.js
