+++ panel/src/ui/index.js
@@ -1,6 +1,7 @@
 /** Plugins */
 import caniuse from "./plugins/caniuse.js";
 import events from "./plugins/events.js";
+import helpers from "./plugins/helpers.js";
 import libraries from "./plugins/libraries.js";
 
 export default {
@@ -12,6 +13,7 @@ export default {
     /** Plugins */
     Vue.use(caniuse);
     Vue.use(events);
+    Vue.use(helpers);
     Vue.use(libraries);
   }
 };
